pyproject.toml
[tool.poetry]
name = "freeprise-matchmaker"
version = "0.1.0"
description = ""
authors = ["BOUSAROUT <mohamed.bousarout@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "1.3.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
